requirements.txt
streamlit>=1.30
numpy
pandas
scikit-learn
umap-learn
plotly
scipy
kneed
matplotlib
